+++ src/main/java/com/hobby/chain/like/service/LikeService.java
@@ -3,5 +3,5 @@ package com.hobby.chain.like.service;
 public interface LikeService {
     public void like(long postId);
     public void unlike(long postId);
-    public int getLikeCount(long postId);
+    public long getLikeCount(long postId);
 }
